requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
boto3==1.28.57
Werkzeug==2.3.7
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-storage==2.10.0
insightface
onnxruntime-silicon
zeroconf>=0.132.2
